chore(.luacheckrc): update globals
Add `Unit` to `.luacheckrc`.

Given limitations in Pico-8's Lua environment, it does not appear as if
using Lua's `require` method for module loading will be possible. That
being the case, the plan going forward will be to add classnames to
`.luacheckrc` as necessary in order to appease the linter.